release 0.9.0
- feat(bind): add more RR types to zoneRR pattern
- style: move class ZONE from ./index to lib/zone
- test: replace coverage reporter nyc with c8
- bind: when a RR doesn't parse, show it before the error
