+++ .github/workflows/master.yml
@@ -41,8 +41,8 @@ jobs:
 
       - name: Coverage Run
         run: |
-          npm install --no-save nyc codecov
-          npx nyc --reporter=lcovonly npm run test
+          npm install --no-save c8 codecov
+          npx c8 --reporter=lcovonly npm run test
 
       - name: Coveralls
         uses: coverallsapp/github-action@master
