+++ lib/bind.js
@@ -14,7 +14,7 @@ export default { zoneOpts, parseZoneFile }
 const re = {
   zoneTTL   : /^\$TTL\s+([0-9]{1,5})\s*(;.*)?$/,
   zoneOrigin: /^\$ORIGIN\s+([^\s]+?)\s*(;.*)?$/,
-  zoneRR    : /^([^\s]+)?\s*([0-9]{1,5})?\s*(IN|CS|CH|HS|NONE|ANY)?\s+(A|AAAA|CAA|CNAME|DNAME|DNSKEY|DS|HINFO|LOC|MX|NAPTR|NS|NSEC|NSEC3|PTR|RRSIG|SMIMEA|SSHFP|SOA|SPF|SRV|TLSA|TXT|URI|TYPE)\s+(.*?)\s*$/,
+  zoneRR    : /^([^\s]+)?\s*([0-9]{1,5})?\s*(IN|CS|CH|HS|NONE|ANY)?\s+(A|AAAA|APL|CAA|CERT|CNAME|DHCID|DNAME|DNSKEY|DS|HINFO|HIP|IPSECKEY|KEY|KX|LOC|MD|MF|MX|NAPTR|NS|NSEC|NSEC3|NSEC3PARAM|NXT|OPENPGPKEY|PTR|RP|RRSIG|SIG|SMIMEA|SOA|SPF|SRV|SSHFP|SVCB|TLSA|TXT|URI|TYPE)\s+(.*?)\s*$/,
   blank     : /^\s*?$/,
   comment   : /^\s*(?:\/\/|;)[^\r\n]*?$/,
 }
@@ -148,15 +148,18 @@ function isZoneOrigin (str, res) {
 }
 
 function parseRR (rr) {
-  switch (rr.type) {
-    case 'SOA'   : return parseSOA(rr)
-    default:
-      return parseAny(rr.type, rr)
+  try {
+    switch (rr.type) {
+      case 'SOA': return parseSOA(rr)
+      default:
+        return new RR[rr.type]({ bindline: `${rr.owner} ${rr.ttl} ${rr.class} ${rr.type} ${rr.rdata}` })
+    }
   }
-}
-
-function parseAny (type, rri) {
-  return new RR[type]({ bindline: `${rri.owner} ${rri.ttl} ${rri.class} ${rri.type} ${rri.rdata}` })
+  catch (e) {
+    console.error(rr)
+    console.error(e.message)
+    throw e
+  }
 }
 
 function parseSOA (rri) {
@@ -178,7 +181,7 @@ function parseSOA (rri) {
   return rrsoa
 }
 
-// TODO: integrate this with parseZone
+// TODO: integrate these remnants with parseZone
 export async function expandShortcuts (zoneArray) {
 
   const implicitOrigin = rr.fullyQualify(zoneOpts.origin) // zone 'name' in named.conf
