+++ lib/tinydns.js
@@ -3,6 +3,8 @@ import * as RR from 'dns-resource-record'
 
 const rr = new RR.A(null)
 
+export default { parseData }
+
 export async function parseData (str) {
   // https://cr.yp.to/djbdns/tinydns-data.html
   const rrs = []
