+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "dns-zone-validator",
-  "version": "0.8.5",
+  "version": "0.9.0",
   "description": "DNS Zone",
   "main": "index.js",
   "type": "module",
